butano: add config `BN_CFG_SPRITE_TEXT_MAX_UTF8_CHARACTERS_FOR_DUPLICATION_CHECK` to save compile time when building fonts with big charaset
Current implementation of _duplicated_utf8_characters() is inefficient
with time complexity of O(N^2). Project build process will hang for an untolerable long time when many characters exist.The duplication check for fonts with big
charaset can be done in font making or converting phase. It is
unnecessary to do it in assertion.